Update config to reflect bulk-redirect@0.2.0
This version of bulk-redirect plugin allows specifying a `basepath`
which is used to resolve URLs values given for the `to` properties in
the `redirects` array. The value of `basepath` is passed as the first
argument to Node's `url.resolve()` API, as `url.resolve(basepath, to)`.

In our case, this generates stub HTML files at the path specified in the
`from` keys, relative to the output root.

Let us consider this example:

```
"bulk-redirect": {
  "basepath": "/documentation/user/current/",
  "redirects": [
    { "from": "configuration/index.html",
      "to": "advanced_readings/configuration/index.html" }
  ]
}
```

In this scenario, if the output location of the book is `_book`, a stub
HTML file will be created at the path `_book/configuration/index.html`,
which will have a redirection to the URL
`/documentation/user/current/advanced_readings/configuration/index.html`.

The value in `from` is used to create the file relative to the output
directory. The value in `to` is resolved with the value of `basepath`,
and added to content of the file created as the URL to redirect to. So,
even if you move the generated file around, it will still point to the
same generated URL.

In our case, when we put the value of `/documentation/user/current/` as
`basepath`, it will solve all breaking URLs that we have currently. But,
when we archive the current documentation, say as v0.6.0, the
redirection URLs in v0.6.0 will still point to
`/documentation/user/current/` because these are statically generated
files, and you cannot change the values once they are generated.

If this is an undesirable effect, then the only way out is to go back to
relative URLs, where the value of `to` has to be manually calculated as
the value relative to the location of the basename of the `from`
value. In that case, the value of `basepath` will need to changed to
`.` (a dot).

Signed-off-by: Kaustav Das Modak <kaustav.dasmodak@yahoo.co.in>